requirements.txt
Django==4.0.5
djangorestframework==3.13.1
psycopg2-binary==2.9.3
Pillow==9.1.1
python-decouple
django-filter
djangorestframework_simplejwt
python-slugify
drf-yasg